+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.0 (in development)
 - Add state when a candidate was contacted and we wait his feedback for interview planning
+- Add view to import seekube ics file. Need to add a settings to your local.py file SEEKUBE_SOURCE_ID parameter
 
 # 1.1.1 (2019-09-09)
 Fix python 3.5 support